Added several fixes to subset() 1) Can now accept numeric concatenated ranges 1//2//3//4//5 2) Strings like "1//2//3//4//5", now work as $2 3) Added a fix to allow variables to be passed as x=,y=,z=
This is a weird bug as it looks like the evaluation of x!="," links all the way back up to a variable from the command line and davinci complains that the variable doesn't exist. This definitely warrants more investigation in the future.  It was mitigated by setting the pass by name variables to new variables in the script.  Which seems to work just fine.