Fixed a bug in sma() that would not allow nn=1 and surface=1 and would return bad data.
Basically, it comes down to the an assumption of what the variable "npositive" actually meant.  It was assumed that it represented the number of endmembers that ARE positive, where it actually means the number of endmembers that CAN ONLY be positive (e.g. the count of endmembers in $2).

These options now provide correct output