Refactored/cleaned up ff_source.c etc.
Rolled the 3 manual separate stacks into a single cvector_void of a
structure since it was really 1 stack of 3 bits of info that went
together.

I added functions to access the relevant info so callers wouldn't have
to do access the vector directly with extern and I can make it static to
ff_source.c.