Fix a segfault that's been there since 1999
A line like any of the following would cause a segfault
printf("%f")
printf("%f", atof(not_a_string))

or as I discovered it in the previous commits
printf("\nDavinci Core Version:\t%.2f\n", atof(version_string()[18:]))

where version_string was old version expecting the old, longer version
string.  So [18:] caused "Invalid range" which caused atof to "expected
a string" and printf got a NULL instead of a float which, finally,
due to this incorrectly retured value, cauled a crash.

This is why we should stick to either 0 or 1 as the error code as much
as possible and not mix.  I think 0 is better as it's the false/NULL
value and thus matches the *alloc functions and makes sense for
predicate functions like if (is_some_property()) even though that's
technically not an "error" just a status.