Added support for reading 1-8 byte int types in unpack
So now it can handle U1-8u1-8I1-8i1-8 and it'll store in the appropriate
8, 16, 32 or 64 bit int or uint type correctly.

I created a test file test64bit.dat file.  Working as expected.  Need
to add it to unpack.dvtest.

Also should create some hdf files of of the unpack test files
with an older version of davinci and make sure the new version
reads exactly the same.  99.9% sure it does.