This doesn't work because ...
I can't delete the left hand side before evaluating the right hand side
if the right hand side may use the existing value of the left hand side.

D'oh!  20/20 hindsight.  Other than that it'd work perfectly fine.

Now the only way I can think to do it is a terrible, horrible hack using
a global variable that is assigned the the lhs which, on a case by case
basis, inside a function, after all the argument parsing/capture, you
could free/remove from the symbol table.

Except even that doesn't work because you could have something like this

A = something
A = func1(A, func2(100, A, blah), other_arg)

and there's no way for func2 to know that it can't delete A yet because
func1 still needs it.

So all this thought/experimentation is for naught except for the
inconsistent returning NULL behavior which I've fixed.