Fixed the leak, removed unused code, ...
I'm so stupid about that leak ... but the election was yesterday so that
was kind of distracting and I was up till 4am watching it last night
which explains my slowness today.  yeah that's my excuse.

Also commented about symbol.c:ff_delete() and noticed ff.c:ff_deleted().

Both totally undocumented and not used anywhere in the library.